// apps/@meetings/app/src/features/audio-upload/model/types.ts
// Recidiviz - a data platform for criminal justice reform
// Copyright (C) 2026 Recidiviz, Inc.
//
// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU General Public License as published by
// the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU General Public License for more details.
//
// You should have received a copy of the GNU General Public License
// along with this program.  If not, see <https://www.gnu.org/licenses/>.
// =============================================================================

export type AudioUploadStatus =
  | "selecting"
  | "uploading"
  | "uploaded"
  | "confirming-error"
  | null;

export type AudioUploadDialog = "cancel" | "error" | "success" | null;

export type RawFileInfo = {
  uri: string;
  name: string;
  mimeType?: string;
  size?: number;
};

export type FileInfo = {
  uri: string;
  name: string;
  contentType: string;
  extension: string;
  size: number;
};